requirements.txt
langchain
langchain-community
langchain-openai
langchain-chroma
langchain-text-splitters
openai
pypdf
pymupdf
fastapi
uvicorn
streamlit
python-dotenv
rank_bm25
ragas
pydantic
unstructured
langchain-google-genai
pypdf
